requirements.txt
numpy==2.0.2
matplotlib==3.10.0
pandas==2.2.2
pyyaml==6.0.3